fix(rooch-da): improve tx_order comparison and error handling
Simplifies the `binary_search_by` logic and refines error handling for `tx_order` mismatch. Ensures alignment with expected behavior and improves code readability.

+++ crates/rooch/src/commands/da/commands/mod.rs
@@ -218,7 +218,7 @@ impl TxOrderHashBlockGetter {
     ) -> anyhow::Result<Vec<TxOrderHashBlock>> {
         let r = self
             .tx_order_hash_blocks
-            .binary_search_by(|x| x.tx_order.cmp(&(start_tx_order as u64)));
+            .binary_search_by(|x| x.tx_order.cmp(&start_tx_order));
         let start_idx = match r {
             Ok(i) => i,
             Err(_) => {
